feat: metrics API revision based on specs and rfcs
