+++ packages/opentelemetry-types/src/index.ts
@@ -19,11 +19,10 @@ export * from './context/propagation/BinaryFormat';
 export * from './context/propagation/HttpTextFormat';
 export * from './distributed_context/DistributedContext';
 export * from './distributed_context/EntryValue';
-export * from './metrics/counter';
-export * from './metrics/gauge';
-export * from './metrics/measure';
-export * from './metrics/meter';
-export * from './metrics/metrics';
+export * from './metrics/Handle';
+export * from './metrics/Measure';
+export * from './metrics/Meter';
+export * from './metrics/Metric';
 export * from './resources/Resource';
 export * from './trace/attributes';
 export * from './trace/Event';
