+++ packages/opentelemetry-types/src/metrics/Meter.ts
@@ -14,30 +14,31 @@
  * limitations under the License.
  */
 
-import { SpanContext } from '../trace/span_context';
-import { DistributedContext } from '../distributed_context/DistributedContext';
-import { Measure, MeasureOptions, Measurement } from './measure';
-import { Metric, MetricOptions } from './metric';
-import { CounterTimeseries } from './counter';
-import { GaugeTimeseries } from './gauge';
+import { Metric, MetricOptions } from './Metric';
+import { CounterHandle, GaugeHandle } from './Handle';
+import { MeasureOptions, Measure } from './Measure';
 
-export interface RecordOptions {
-  // spanContext represents a measurement exemplar in the form of a SpanContext.
-  spanContext?: SpanContext;
-  // distributedContext overrides the current context and adds dimensions
-  // to the measurements.
-  distributedContext?: DistributedContext;
-}
-
+/**
+ * An interface to allow the recording metrics.
+ *
+ * {@link Metric}s are used for recording pre-defined aggregation (Gauge and
+ * Counter), or raw values ({@link Measure}) in which the aggregation and labels
+ * for the exported metric are deferred.
+ */
 export interface Meter {
-  // Creates and returns a new @link{Measure}.
+  /**
+   * Creates and returns a new {@link Measure}.
+   * @param name the name of the metric.
+   * @param [options] the measure options.
+   */
   createMeasure(name: string, options?: MeasureOptions): Measure;
 
-  // Creates a new counter metric.
-  createCounter(
-    name: string,
-    options?: MetricOptions
-  ): Metric<CounterTimeseries>;
+  /**
+   * Creates a new counter metric.
+   * @param name the name of the metric.
+   * @param [options] the metric options.
+   */
+  createCounter(name: string, options?: MetricOptions): Metric<CounterHandle>;
 
   // TODO: Measurements can have a long or double type. However, it looks like
   // the metric timeseries API (according to spec) accepts values instead of
@@ -47,9 +48,10 @@ export interface Meter {
   // be cool to only have a single interface, but maybe having two is necessary?
   // Maybe include the type as a metrics option? Probs a good gh issue, the same goes for Measure types.
 
-  // Creates a new gauge metric.
-  createGauge(name: string, options?: MetricOptions): Metric<GaugeTimeseries>;
-
-  // Record a set of raw measurements.
-  record(measurements: Measurement[], options?: RecordOptions): void;
+  /**
+   * Creates a new gauge metric.
+   * @param name the name of the metric.
+   * @param [options] the metric options.
+   */
+  createGauge(name: string, options?: MetricOptions): Metric<GaugeHandle>;
 }
