--- packages/opentelemetry-types/src/metrics/gauge.ts
@@ -1,22 +1,0 @@
-/**
- * Copyright 2019, OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export interface GaugeTimeseries {
-  // Adds the given value to the current value. Values can be negative.
-  add(value: number): void;
-  // Sets the given value. Values can be negative.
-  set(value: number): void;
-}
